Fix cyface sync android tests
