+++ datacapturing/src/main/java/de/cyface/datacapturing/MessageCodes.java
@@ -13,7 +13,7 @@ import de.cyface.synchronization.SyncService;
  *
  * @author Klemens Muthmann
  * @author Armin Schnabel
- * @version 1.1.3
+ * @version 1.1.4
  * @since 2.0.0
  */
 public class MessageCodes {
