+++ synchronization/src/androidTest/java/de/cyface/synchronization/TestUtils.java
@@ -15,7 +15,8 @@ import de.cyface.persistence.RotationPointTable;
  * Contains utility methods and constants required by the tests within the synchronization project.
  *
  * @author Klemens Muthmann
- * @version 1.0.4
+ * @author Armin Schnabel
+ * @version 1.0.5
  * @since 2.1.0
  */
 final class TestUtils {
@@ -35,11 +36,16 @@ final class TestUtils {
     /**
      * An username used by the tests to set up a Cyface account for synchronization.
      */
-    public final static String DEFAULT_FREE_USERNAME = "playStoreBeta";
+    public final static String DEFAULT_USERNAME = "admin";
     /**
      * A password used by the tests to set up a Cyface account for synchronization.
      */
-    public final static String DEFAULT_FREE_PASSWORD = "playStoreBeta@Cy";
+    public final static String DEFAULT_PASSWORD = "secret";
+
+    /**
+     * Path to an API available for testing.
+     */
+    public final static String TEST_API_URL = "https://192.168.1.146:8080/api/v2";
 
     static Uri getMeasurementUri() {
         return new Uri.Builder().scheme("content").authority(AUTHORITY).appendPath(MeasurementTable.URI_PATH).build();
