+++ synchronization/src/androidTestMovebis/java/de/cyface/synchronization/DataTransmissionTest.java
