+++ synchronization/src/cyface/res/xml/authenticator.xml
@@ -1,5 +1,5 @@
 <account-authenticator xmlns:android="http://schemas.android.com/apk/res/android"
-                       android:accountType="de.cyface.synchronization.test"
+                       android:accountType="de.cyface.account"
      android:icon="@drawable/ic_launcher_foreground"
      android:smallIcon="@drawable/ic_launcher_foreground"
      android:label="@string/cyface"/>
